Add enumerate_repo method to distro.pkg_manager
Takes repo path, finds all packages contained in it.

For deb: find the Origin tag in the repo, and use aptitude to
search for all packages from that origin.

For rpm: find all the rpm files in the path, and use rpm -qp
with a --queryformat to parse out the package name.

Signed-off-by: Dan Mick <dan.mick@inktank.com>