Implement homegrown strtobool
Create a simpler strtobool that just looks for yes/no combinations
rather than also allowing numbers ('1', '0') or things like "on"
or "off".  This allows us to remove the previous PSFL code.

Make sure that we maintain the previous behavior where not entering
a response to the prompt evaluates to True.  True is the default
value for an empty string.

Update tests to account for new behavior.

Signed-off-by: Travis Rhoden <trhoden@redhat.com>