Add -d/--dir option to ice_setup
This enhances the ice_setup script to accept a -d/--dir option
that contains the path to where the script should look for package
files to add to the locall hosted repos.

Previously, the script only looked in the same directory as the
script (which may not even be the CWD). To preserve backwads
compatibility, --dir is optional.  If not provided, the default path
is now the CWD.  When run without any subcommands (i.e. 'ice-setup'),
the interactive prompts show the user the path to search, and allow
the user to modify it.  providing --dir to the script merely changes
this default.

This is a big patch that was difficult to split into lots of smaller
ones.  I found a lot of the terminology confusing -- referring to the
sources as "repos" when they are not actually repos.  They are folders
of package files that should be used to create a formal repo. I've
changed many of the function and variable names to reflect that.

Signed-off-by: Travis Rhoden <trhoden@redhat.com>