Fix behavior of 'ice_setup configure ...'
The Configure subcommands are there to allow a user to update/copy
new package files into the ICE node repositories, without going
through the interactive prompts. The user can pick the repos
for local use (calamari-server, ceph-deploy) or remote use (ceph,
calamari-clients), or all.

This functionality was broken previously because the calls to
configure_[local,remote] were missing the needed names.

The configure commands already accepted a path to grab the package files
from, but the path is optional. no dash or double-dash is needed for the
path, and if not given then CWD is used for the path.

Signed-off-by: Travis Rhoden <trhoden@redhat.com>