Improve error-handling in get_package_source
Enhance get_package_source to know the difference between a top-level
folder not existing (i.e. 'ceph') and the versioned folder when it is
expected (i.e. 'ceph/0.80'). Explicity check that the paths are
directories, not just the path exists.

Since os.walk is returning a list of directories, and we are using an
entry from that list, there is no reason to check that it is a directory
once again.

Signed-off-by: Travis Rhoden <trhoden@redhat.com>