clarify HTTPS prompt
If a user selects "https" at the prompt, we require that the user
has pre-configured Apache to serve HTTPS.

Prior to this commit, it was not clear to users that they must set up
HTTPS by hand prior to running ice_setup.

Adjust the prompt so that it is clear that HTTPS setup must be done
manually, and place a stronger emphasis on choosing the "http" default.

See https://bugzilla.redhat.com/1197883 for the original report.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>