require pkg_resources module at runtime
The /usr/bin/ice_setup utility uses load_entry_point from the
pkg_resources module. If this module is not available,
/usr/bin/ice_setup will crash.

Require the OS packages that include pkg_resources so that this
dependency is available at runtime.

There were some scenarios where the python-setuptools package (on Red
Hat) or python-pkg-resources package (on Ubuntu) was unavailable (for
example, in minimal OS installs).

Refs: https://bugzilla.redhat.com/1212045

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>