ice.py: better errors when writing cephdeploy.conf
ice_setup doesn't check to see if CWD is writable before writing to
os.path.join(CWD, 'cephdeploy.conf'). This can lead to a stack trace if
CWD is not writable. Users commonly encounter this when running
ice_setup while cd'd to the RHCS ISO mount path (/mnt), since that
directory is read-only.

When we encounter a failure to write cephdeploy.conf, catch the IOError.
If it's an EROFS error, print a friently message to the user. In any
case, just print the error string itself, without printing a full stack
trace.

(A longer-term fix in ice_setup itself would involve not writing so many
cephdeploy.conf files.)